Cell splitting in Maya
This is one of orb behaviors - splitting one orb into two.

Idea: 
A Merge cube generates an orb; when the scene starts, there's only one orb, but slowly, the orb multiplies into two, and then two into four... The goal of it is to achieve a hypnotizing effect when users focus on the slow motions of the orbs. 

Challenge: 
We couldn't export the Maya animation into Unity, because Maya's "blend shape" feature is not supported by Unity, but we are putting it here for records!